Document accessing form data
Form values in PDF files are handled differently than other text, so
it's not supported through the existing pdfminer APIs, but it can be
access through pdfminer's wrapping of pdfplumber.

Add a section to the README to save others the confusion I faced when
figuring out how to access form data.

Addresses #120
https://github.com/jsvine/pdfplumber/issues/120