Fix slowdown in extract_words on long words (#483)
This commit fixes the slowdown in `.extract_words(...)` and
`WordExtractor.iter_chars_to_words(...)` which occurred for very long
"words", and which was caused by repeatedly re-calculating bounding box.

See #483 for discussion and example.

It also adds `utils.merge_bboxes(bboxes)`, which returns the smallest
bounding box that contains all bounding boxes in the `bboxes` argument,
and which we use in the fix.